+++ src/tests.rs
@@ -15,8 +15,10 @@ fn check_struct_sizes() {
 
     assert_eq!(size_of::<Dynamic>(), if PACKED { 8 } else { 16 });
     assert_eq!(size_of::<Option<Dynamic>>(), if PACKED { 8 } else { 16 });
-    #[cfg(not(feature = "no_position"))]
-    assert_eq!(size_of::<Position>(), 4);
+    assert_eq!(
+        size_of::<Position>(),
+        if cfg!(feature = "no_position") { 0 } else { 4 }
+    );
     assert_eq!(size_of::<ast::Expr>(), 16);
     assert_eq!(size_of::<Option<ast::Expr>>(), 16);
     assert_eq!(size_of::<ast::Stmt>(), 32);